+++ Source/Podio .NET/Services/ReminderService.cs
@@ -17,7 +17,7 @@ namespace PodioAPI.Services
         /// <param name="refType"></param>
         /// <param name="refId"></param>
         /// <returns></returns>
-        public async Task<int> GetReminder(string refType, int refId)
+        public async Task<int> GetReminder(string refType, long refId)
         {
             string url = string.Format("/reminder/{0}/{1}", refType, refId);
             dynamic response =  await _podio.Get<dynamic>(url);
@@ -31,7 +31,7 @@ namespace PodioAPI.Services
         /// <param name="refType"></param>
         /// <param name="refId"></param>
         /// <param name="reminderId"></param>
-        public async Task<dynamic> DeleteReminder(string refType, int refId, int reminderId)
+        public async Task<dynamic> DeleteReminder(string refType, long refId, int reminderId)
         {
             string url = string.Format("/reminder/{0}/{1}?reminder_id={2}", refType, refId, reminderId);
             return  await _podio.Delete<dynamic>(url);
@@ -44,7 +44,7 @@ namespace PodioAPI.Services
         /// <param name="refType"></param>
         /// <param name="refId"></param>
         /// <param name="reminderId"></param>
-        public async Task<dynamic> SnoozeReminder(string refType, int refId, int reminderId)
+        public async Task<dynamic> SnoozeReminder(string refType, long refId, int reminderId)
         {
             string url = string.Format("/reminder/{0}/{1}/snooze?reminder_id={2}", refType, refId, reminderId);
             return  await _podio.Post<dynamic>(url);
@@ -57,7 +57,7 @@ namespace PodioAPI.Services
         /// <param name="refType"></param>
         /// <param name="refId"></param>
         /// <param name="remindDelta">minutes to remind before the due date of the object</param>
-        public async Task<dynamic> Update(string refType, int refId, int remindDelta)
+        public async Task<dynamic> Update(string refType, long refId, int remindDelta)
         {
             string url = string.Format("/reminder/{0}/{1}", refType, refId);
             dynamic requestData = new
